Cast Spell menu color coding and filter.
Spells with 0% chance of success, are now hidden from this menu. You can still see them in the menus to view, mix, manage and reorder spells.
I also color-coded the spell name. If less than 50% chance of success (but more than 0%), it's yellow.
If there is one casting left, or you have enough mana for one casting, it's orange.
If you don't have enough mana to cast it, or you have zero castings left, it's red.